feature/initial-page: changed the initial page to the right place
